move CountRope
